requirements.txt
pandas
pony
gensim
scikit-learn
imbalanced-learn 
requests
bs4
pyenchant

